# crates/aivyx-server/src/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Aivyx Engine API",
    "version": "1.1.0",
    "description": "AI agent orchestration engine with multi-agent teams, A2A/MCP protocol support, and enterprise features.",
    "license": {
      "name": "Proprietary"
    }
  },
  "servers": [
    { "url": "http://localhost:3000", "description": "Local development" },
    { "url": "https://api.aivyx.ai", "description": "Production" }
  ],
...
  "paths": {
...
    "/plugins": {
      "get": {
        "summary": "List plugins",
        "tags": ["Plugins"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "List of installed plugins",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "type": "object" }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Install plugin",
        "tags": ["Plugins"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "source": { "type": "string" }
                },
                "required": ["name"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Plugin installed" }
        }
      }
    },
    "/plugins/{name}": {
      "delete": {
        "summary": "Remove plugin",
        "tags": ["Plugins"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "description": "Plugin name"
          }
        ],
        "responses": {
          "204": { "description": "Plugin removed" }
        }
      }
    },
...
    "/skills": {
      "get": {
        "summary": "List skills",
        "tags": ["Skills"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "List of skills",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "type": "object" }
                }
              }
            }
          }
        }
      }
    },
    "/skills/effectiveness": {
      "get": {
        "summary": "Skill effectiveness scores",
        "tags": ["Skills"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Effectiveness scores for all skills",
            "content": {
              "application/json": {
                "schema": { "type": "object" }
              }
            }
          }
        }
      }
    },
    "/skills/{name}": {
      "get": {
        "summary": "Get skill",
        "tags": ["Skills"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "description": "Skill name"
          }
        ],
        "responses": {
          "200": {
            "description": "Skill details",
            "content": {
              "application/json": {
                "schema": { "type": "object" }
              }
            }
          },
          "404": {
            "description": "Skill not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Error" }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete skill",
        "tags": ["Skills"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "description": "Skill name"
          }
        ],
        "responses": {
          "204": { "description": "Skill deleted" }
        }
      }
    },
...
    "/secrets": {
      "get": {
        "summary": "List secrets",
        "description": "List secret names (values are never returned).",
        "tags": ["Secrets"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "List of secret names",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "type": "string" }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Set secret",
        "tags": ["Secrets"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "value": { "type": "string" }
                },
                "required": ["name", "value"]
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Secret stored" }
        }
      }
    },
    "/secrets/{name}": {
      "delete": {
        "summary": "Delete secret",
        "tags": ["Secrets"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": { "type": "string" },
            "description": "Secret name"
          }
        ],
        "responses": {
          "204": { "description": "Secret deleted" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Bearer token or multi-tenant API key"
      }
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string", "description": "Error code" },
          "message": { "type": "string", "description": "Human-readable error message" }
        },
        "required": ["error", "message"]
      }
    }
  },
  "security": [
    { "bearerAuth": [] }
  ]
}